+++ src/doubleLinkedL_Vorlesung/ReverseIterator.java
@@ -1,4 +1,4 @@
-package woche10;
+package doubleLinkedL_Vorlesung;
 
 public class ReverseIterator {
