+++ src/fraction_testat/Testumgebung.java
@@ -1,4 +1,4 @@
-package woche5;
+package fraction_testat;
 
 public class Testumgebung {
