+++ src/ub5_praktikum3/Ubungsblatt5.java
@@ -1,4 +1,4 @@
-package woche6;
+package ub5_praktikum3;
 
 import extra.Tools;
